refactor: reorganise docker command runner to be more generic
