--- topo_processor/cog/__init__.py
@@ -1,2 +1,0 @@
-from .command import Command
-from .execution import ExecutionDocker, ExecutionLocal
